+++ src/pages/articles/[articleId]/index.tsx
@@ -11,16 +11,15 @@ import Skeleton from 'react-loading-skeleton';
 import { Article } from '~/components/Article';
 import { ArticleError } from '~/components/Article/ArticleError';
 import ArticleComment from '~/components/ArticleComment';
-import {
-  FullPageLoader,
-  PageHead,
-  PageHeadingText,
-} from '~/components/Common';
+import { FullPageLoader } from '~/components/Common/FullPageLoader';
+import { PageHead } from '~/components/Common/PageHead';
+import { PageHeadingText } from '~/components/Common/PageHeadingText';
 import ArticleCommentForm from '~/components/Forms/ArticleCommentForm';
 import TitleBar from '~/components/TitleBar';
 import { queryKeys } from '~/react-query/common';
 import { prefetch } from '~/react-query/server';
-import { getArticleDetail, useArticleDetail } from '~/services/article';
+import { getArticleDetail } from '~/services/article/apis';
+import { useArticleDetail } from '~/services/article/hooks';
 import {
   useArticleComments,
   useCreateArticleComment,
@@ -34,7 +33,8 @@ import {
   palettes,
   titleBarHeight,
 } from '~/styles/utils';
-import { handleAxiosError, routes } from '~/utils';
+import { handleAxiosError } from '~/utils/handleAxiosError';
+import { routes } from '~/utils/routes';
 import { stripHtmlTags } from '~/utils/stripHtmlTags';
 
 interface ArticleDetailPageProps
